+++ src/main/java/com/kingguanzhang/dealsites/service/impl/FavoriteShopServiceImpl.java
@@ -23,10 +23,18 @@ public class FavoriteShopServiceImpl implements FavoriteShopService {
      */
     @Override
     public Integer addFavoriteShop(Integer productId, Integer userId) {
+        if (null == productId || null== userId){
+            throw new RuntimeException("未能获取到商品Id或用户Id");
+        }
         FavoriteShop favoriteShop = new FavoriteShop();
         favoriteShop.setShopId(productId);
         favoriteShop.setPersoninfoId(userId);
-        int i = favoriteShopMapper.insertSelective(favoriteShop);
+        int i=0;
+        try{
+             i = favoriteShopMapper.insertSelective(favoriteShop);
+        }catch (Exception e){
+            throw new RuntimeException("收藏店铺失败");
+        }
         return i;
     }
 
@@ -38,9 +46,17 @@ public class FavoriteShopServiceImpl implements FavoriteShopService {
      */
     @Override
     public Integer removeFavoriteShop(Integer productId, Integer userId) {
+        if (null == productId || null== userId){
+            throw new RuntimeException("未能获取到商品Id或用户Id");
+        }
         FavoriteShopExample favoriteShopExample = new FavoriteShopExample();
         favoriteShopExample.createCriteria().andShopIdEqualTo(productId).andPersoninfoIdEqualTo(userId);
-        int i = favoriteShopMapper.deleteByExample(favoriteShopExample);
+        int i=0;
+        try{
+            i = favoriteShopMapper.deleteByExample(favoriteShopExample);
+        }catch (Exception e){
+            throw new RuntimeException("取消收藏失败");
+        }
         return i;
     }
 
@@ -51,6 +67,9 @@ public class FavoriteShopServiceImpl implements FavoriteShopService {
      */
     @Override
     public List<FavoriteShop> getFavoriteShopList(Integer personInfoId) {
+        if (null == personInfoId ){
+            throw new RuntimeException("未能获取到用户Id");
+        }
         FavoriteShopExample favoriteShopExample = new FavoriteShopExample();
         favoriteShopExample.createCriteria().andPersoninfoIdEqualTo(personInfoId);
         List<FavoriteShop> favoriteShopList = favoriteShopMapper.selectByExample(favoriteShopExample);
@@ -65,6 +84,9 @@ public class FavoriteShopServiceImpl implements FavoriteShopService {
      */
     @Override
     public Integer getFavoriteShop(Integer userId, Integer shopId) {
+        if (null == shopId || null == userId ){
+            throw new RuntimeException("未能获取到用户Id或店铺Id");
+        }
         FavoriteShopExample favoriteShopExample = new FavoriteShopExample();
         favoriteShopExample.createCriteria().andPersoninfoIdEqualTo(userId).andShopIdEqualTo(shopId);
         List<FavoriteShop> favoriteShopList = favoriteShopMapper.selectByExample(favoriteShopExample);
