+++ src/main/java/com/kingguanzhang/dealsites/service/impl/ProductCategoryServiceImpl.java
@@ -21,8 +21,10 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
     private ProductMapper productMapper;
 
 
-
-
+    /**
+     * 获取分类列表;
+     * @return
+     */
     @Override
     public List<ProductCategory> getCategory() {
         //这里需要按照priority排序,这是为了方便在首页的分类列表进行划分;
@@ -32,20 +34,63 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
         return productCategoryList;
     }
 
+    /**
+     * 新增分类;
+     * @param productCategory
+     * @return
+     */
     @Override
     public Integer addProductCategroy(ProductCategory productCategory) {
-        return productCategoryMapper.insert(productCategory);
+        if (null == productCategory){
+            throw new RuntimeException("未能获取到分类信息");
+        }
+        int i=0;
+        try{
+            i = productCategoryMapper.insertSelective(productCategory);
+        }catch (Exception e){
+            throw new RuntimeException("新增分类失败");
+        }
+        return i;
     }
 
+    /**
+     * 更新分类
+     * @param productCategory
+     * @return
+     */
     @Override
-    public Integer editProductCategroy(ProductCategory productCategory) {
-        return productCategoryMapper.updateByPrimaryKeySelective(productCategory);
+    public Integer updateProductCategory(ProductCategory productCategory) {
+        if (null == productCategory){
+            throw new RuntimeException("未能获取到分类信息");
+        }
+        int i=0;
+        try{
+           i = productCategoryMapper.updateByPrimaryKeySelective(productCategory);
+        }catch (Exception e){
+            throw new RuntimeException("更新分类失败");
+        }
+        return i;
     }
 
+    /**
+     * 删除分类;
+     * @param id
+     * @return
+     */
     @Override
     public Integer deleteProductCategroy(Integer id) {
-        //删除分类之前先将分类关联的所有商品的分类设置成NULL;
-        productMapper.updateProductCategoryIdToNullByProductCategoryId(id);
-        return productCategoryMapper.deleteByPrimaryKey(id);
+        if (null == id){
+            throw new RuntimeException("未能获取到分类Id");
+        }
+        int i=0;
+        try{
+            //删除分类之前先将分类关联的所有商品的分类设置成NULL;
+            productMapper.updateProductCategoryIdToNullByProductCategoryId(id);
+          i =  productCategoryMapper.deleteByPrimaryKey(id);
+        }catch (Exception e){
+            throw new RuntimeException("删除分类失败");
+        }
+
+        return i;
     }
 }
